Refactor chat list manager to package database.
<h1>Rationale:</h1>

This commit moves and renames ChatListManager to package database as DatabaseListManager.  No functional, cosmetic or other changes are included.  Pure and simple refactor.  This is being done to improve the ability to retrofit upcoming changes for adding experiences to chats.
